+++ CHANGELOG.md
@@ -2,6 +2,41 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [6.17.0](https://github.com/lifinance/data-types/compare/v6.15.0...v6.17.0) (2025-06-11)
+
+
+### Features
+
+* add Flare chain ([#143](https://github.com/lifinance/data-types/issues/143)) ([52f49d7](https://github.com/lifinance/data-types/commit/52f49d79cf8b07a8ad77734b4068bd8151261342))
+* add weth to SOE adn ERA ([#142](https://github.com/lifinance/data-types/issues/142)) ([b5ffc94](https://github.com/lifinance/data-types/commit/b5ffc94e17d0603defd38401632960b85f4cb115))
+
+## [6.16.0](https://github.com/lifinance/data-types/compare/v6.15.0...v6.16.0) (2025-06-11)
+
+
+### Features
+
+* add weth to SOE ([fd19d74](https://github.com/lifinance/data-types/commit/fd19d74e15f49e96ace554fcf24595733b7f47b7))
+
+## [6.15.0](https://github.com/lifinance/data-types/compare/v6.14.6...v6.15.0) (2025-06-05)
+
+
+### Features
+
+* add Viction chain ([#141](https://github.com/lifinance/data-types/issues/141)) ([d9fe606](https://github.com/lifinance/data-types/commit/d9fe6060e948dceccd9a00ff7a60178ab0431b5a))
+
+### [6.14.6](https://github.com/lifinance/data-types/compare/v6.14.5...v6.14.6) (2025-06-05)
+
+
+### Bug Fixes
+
+* **chains:** add etherlink ([#140](https://github.com/lifinance/data-types/issues/140)) ([6791fc0](https://github.com/lifinance/data-types/commit/6791fc0c736c874b51472ab8be28cdc51ade441e))
+
+### [6.14.5](https://github.com/lifinance/data-types/compare/v6.14.4...v6.14.5) (2025-05-30)
+
+### [6.14.4](https://github.com/lifinance/data-types/compare/v6.14.3...v6.14.4) (2025-05-28)
+
+### [6.14.3](https://github.com/lifinance/data-types/compare/v6.14.2...v6.14.3) (2025-05-28)
+
 ### [6.14.2](https://github.com/lifinance/data-types/compare/v6.14.1...v6.14.2) (2025-05-26)
 
 ### [6.14.1](https://github.com/lifinance/data-types/compare/v6.14.0...v6.14.1) (2025-05-23)